maestro-django/webpersonal/requirements.txt
astroid==2.1.0
Django==2.0.2
django-rest-auth==0.9.3
djangorestframework==3.9.0
isort==4.3.4
lazy-object-proxy==1.3.1
mccabe==0.6.1
Pillow==5.3.0
pylint==2.2.2
pytz==2018.7
six==1.11.0
typed-ast==1.1.0
wrapt==1.10.11
